+++ portals/publisher/src/main/webapp/source/src/app/components/GlobalPolicies/Policies/GlobalSpecificComponents/PoliciesSection.tsx
@@ -20,7 +20,7 @@ import { Grid } from '@material-ui/core';
 import React, { FC } from 'react';
 import Box from '@material-ui/core/Box';
 import type { Policy, PolicySpec } from '../Types';
-import PoliciesExpansion from '../PoliciesExpansion';
+import PoliciesExpansion from '../SharedComponents/PoliciesExpansion';
 
 interface PolicySectionProps {
     isChoreoConnectEnabled: boolean;
